# lib/api-zod/src/generated/types/generateResumeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateResumeRequest {
  /** Full name of the candidate */
  name: string;
  /** Comma-separated list of skills */
  skills: string;
  /** Work experience description */
  experience: string;
  /** Education background */
  education: string;
}