# lib/api-zod/src/generated/types/improveResumeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ImproveResumeRequest {
  /** The existing resume text to improve */
  resumeText: string;
}