# lib/api-zod/src/generated/types/interviewRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewRequest {
  /** Job role or position title */
  role: string;
  /** Relevant skills for the role (optional) */
  skills?: string;
}