# lib/api-zod/src/generated/types/interviewResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewQA } from "./interviewQA";

export interface InterviewResponse {
  /** List of interview questions with answers */
  questions: InterviewQA[];
}